requirements.txt
solana
discord